requirements.txt
cvxpy
tensorboard
geopy
#PIL
scikit-learn
networkx
numpy
scipy
matplotlib
jupyter
spacy
mplleaflet
opencv-python
torch_optimizer